Modofied code due to feedback from Mattia Vaccari ert his fusion catalogues:
1) the servs data fusion files you're using are the correct ones.
note that servs positions were *NOT* registered against 2mass,
or rather, the average offsets are in the servs dr1 document but
were not applied to the catalogs, not quite sure why to be honest;/)

2) ra_12 is the average of servs irac1 and irac2 positions,
(when both are available, otherwise it's simply ra_1 or ra_2)
and should be used at all times.

3) aper_2 fluxes are best used rather than aper_1 fluxes.
that can make a difference for slightly extended galaxies.